Make entire `valid::handles` module dependent on "validate" feature.
Rather than having `#[cfg(feature = "validate")]` attributes on almost
every definition in `src/valid/handles.rs`, move the definitions we
need unconditionally out of the module, and just put the whole module
under a `#[cfg]`.
